requirements.txt
# ─── LLM (Apple Silicon Native) ────────────────────────────
mlx-lm>=0.19.0

# ─── Speech-to-Text ─────────────────────────────────────────
openai-whisper>=20231117
sounddevice>=0.4.6
scipy>=1.11.0
numpy>=1.24.0

# ─── Web Server ──────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0

# ─── System Info ─────────────────────────────────────────────
psutil>=5.9.0

# ─── Web Search (DuckDuckGo, no API key needed) ──────────────
duckduckgo-search>=6.0.0

# ─── Utilities ───────────────────────────────────────────────
python-dotenv>=1.0.0
rich>=13.0.0
